% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conversion.R
\name{add_archr_reductions}
\alias{add_archr_reductions}
\title{Add dimensionality reductions from ArchR to Signac}
\usage{
add_archr_reductions(
  archr_proj,
  seurat_obj,
  reductions = c("UMAP", "IterativeLSI"),
  reduction_key = "archr_"
)
}
\arguments{
\item{archr_proj}{ArchRProject object}

\item{seurat_obj}{Seurat object}

\item{reductions}{character vector Names of reductions to add}

\item{reduction_key}{character Key for reduction names}
}
\value{
Updated Seurat object
}
\description{
Add dimensionality reductions from ArchR to Signac
}
