% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_scoring.R
\name{add_element_scores}
\alias{add_element_scores}
\title{Add element scores to ArchR project}
\usage{
add_element_scores(archr_proj, score_df, overwrite = FALSE)
}
\arguments{
\item{archr_proj}{ArchRProject object}

\item{score_df}{data.frame containing element scores}

\item{overwrite}{logical whether to overwrite existing scores}
}
\value{
Updated ArchRProject object
}
\description{
Add element scores to ArchR project
}
\examples{
\dontrun{
updated_proj <- add_element_scores(
  archr_proj = proj,
  score_df = element_scores
)
}
}
