% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_conversion.R
\name{archr_to_signac}
\alias{archr_to_signac}
\title{Convert ArchR project to Signac object}
\usage{
archr_to_signac(
  archr_proj,
  ref_version = c("hg38", "hg19"),
  fragments_dir,
  peak_matrix,
  fragments_from_cellranger = FALSE,
  fragments_extension = NULL,
  annotation
)
}
\arguments{
\item{archr_proj}{ArchRProject object}

\item{ref_version}{character Genome reference version (default: "hg38")}

\item{fragments_dir}{character Directory containing fragment files}

\item{peak_matrix}{matrix Peak matrix from ArchR project}

\item{fragments_from_cellranger}{logical Whether fragments are from cellranger (default: FALSE)}

\item{fragments_extension}{character File extension for fragment files}

\item{annotation}{Object containing gene annotations}
}
\value{
Seurat object with Signac assay
}
\description{
Convert ArchR project to Signac object
}
