% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_scoring.R
\name{calculate_element_scores}
\alias{calculate_element_scores}
\title{Calculate element-level accessibility scores}
\usage{
calculate_element_scores(te_matrix, te_granges)
}
\arguments{
\item{te_matrix}{SummarizedExperiment TE accessibility matrix}

\item{te_granges}{GRanges object containing TE annotations}
}
\value{
data.frame with element-level accessibility scores for each cell
}
\description{
Calculate element-level accessibility scores
}
\examples{
\dontrun{
element_scores <- calculate_element_scores(
  te_matrix = te_mat,
  te_granges = te_loci
)
}
}
