% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_catalog.R
\name{combine_te_catalogs}
\alias{combine_te_catalogs}
\title{Combine multiple TE catalogs}
\usage{
combine_te_catalogs(catalog_list)
}
\arguments{
\item{catalog_list}{list of GRanges objects containing TE annotations}
}
\value{
Combined GRanges object
}
\description{
Combine multiple TE catalogs
}
\examples{
\dontrun{
combined_catalog <- combine_te_catalogs(
  list(l1_catalog, sva_catalog)
)
}
}
