% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{createTESet}
\alias{createTESet}
\title{Create TESet object}
\usage{
createTESet(counts, rowData = NULL, colData = NULL, metadata = list())
}
\arguments{
\item{counts}{matrix Count matrix}

\item{rowData}{DataFrame Row annotations}

\item{colData}{DataFrame Column annotations}

\item{metadata}{list Additional metadata}
}
\value{
TESet object
}
\description{
Create TESet object
}
