% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_scoring.R
\name{filter_te_cells}
\alias{filter_te_cells}
\title{Filter cells based on TE accessibility}
\usage{
filter_te_cells(
  archr_proj,
  matrix_name = "TEMatrix",
  min_counts = NULL,
  min_detected = NULL
)
}
\arguments{
\item{archr_proj}{ArchRProject object}

\item{matrix_name}{character Name of the TE matrix}

\item{min_counts}{numeric Minimum number of counts required}

\item{min_detected}{numeric Minimum number of TEs detected}
}
\value{
Vector of cell names passing filters
}
\description{
Filter cells based on TE accessibility
}
\examples{
\dontrun{
valid_cells <- filter_te_cells(
  archr_proj = scored_proj,
  min_counts = 100,
  min_detected = 10
)
}
}
