% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_catalog.R
\name{filter_te_loci}
\alias{filter_te_loci}
\title{Filter TE loci based on peaks and blacklist}
\usage{
filter_te_loci(
  te_catalog,
  peak_regions = NULL,
  blacklist = NULL,
  upstream_extension = 1000
)
}
\arguments{
\item{te_catalog}{GRanges object containing TE annotations}

\item{peak_regions}{GRanges object containing ATAC-seq peaks}

\item{blacklist}{GRanges object containing blacklist regions}

\item{upstream_extension}{numeric bases to extend upstream for regulatory regions}
}
\value{
GRanges object containing filtered TE annotations
}
\description{
Filter TE loci based on peaks and blacklist
}
\examples{
\dontrun{
filtered_catalog <- filter_te_loci(
  te_catalog = combined_catalog,
  peak_regions = peaks,
  blacklist = blacklist_regions,
  upstream_extension = 1000
)
}
}
