% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_te_coverage}
\alias{plot_te_coverage}
\title{Plot TE coverage using Signac}
\usage{
plot_te_coverage(
  seurat_obj,
  te_regions,
  group_by = NULL,
  extend_upstream = 3000,
  extend_downstream = 3000
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{te_regions}{GRanges object containing TE regions}

\item{group_by}{character Column name for grouping}

\item{extend_upstream}{numeric Bases to extend upstream}

\item{extend_downstream}{numeric Bases to extend downstream}
}
\value{
ggplot object
}
\description{
Plot TE coverage using Signac
}
