% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/te_heterogeneity.R
\name{plot_te_heterogeneity}
\alias{plot_te_heterogeneity}
\title{Plot TE accessibility on dimensional reduction}
\usage{
plot_te_heterogeneity(
  archr_proj,
  matrix_name = "TEMatrix",
  feature_name,
  reduction = "UMAP",
  use_imputation = TRUE
)
}
\arguments{
\item{archr_proj}{ArchRProject object}

\item{matrix_name}{character Name of the matrix (default: "TEMatrix")}

\item{feature_name}{character Name of the TE feature to plot}

\item{reduction}{character Name of dimensionality reduction (default: "UMAP")}

\item{use_imputation}{logical Whether to use imputation (default: TRUE)}
}
\value{
ggplot object
}
\description{
Plot TE accessibility on dimensional reduction
}
\examples{
\dontrun{
p <- plot_te_heterogeneity(
  archr_proj = proj,
  feature_name = "L1HS_chr1_123456_123789_+",
  reduction = "UMAP"
)
}
}
