% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{plot_te_qc}
\alias{plot_te_qc}
\title{Plot QC metrics for TE accessibility}
\usage{
plot_te_qc(archr_proj, matrix_name = "TEMatrix")
}
\arguments{
\item{archr_proj}{ArchRProject object}

\item{matrix_name}{character Name of TE matrix}
}
\value{
List of ggplot objects
}
\description{
Plot QC metrics for TE accessibility
}
