% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_classes.R
\name{show,TESet-method}
\alias{show,TESet-method}
\title{Show method for TESet objects}
\usage{
\S4method{show}{TESet}(object)
}
\arguments{
\item{object}{TESet object}
}
\description{
Show method for TESet objects
}
